Set dock focused icon before zoom down animation   - Switch to SBIconFadeAnimator to handle 'Reduce Motion' setting
When an app icon was launched, it stayed in focus until the app was exited,
and then dock was collapsed. This caused problems when the user switched
apps via the app switcher or a notification, as the icon in focus was
not the correct icon for the app being zoomed down.

To fix this, the dock is now collapsed after an app icon is launched and the
animation is complete. When an app is exited, the dock's focused icon is
set to that app's icon, assuming the icon is in the dock.